+++ NEWS.md
@@ -3,8 +3,10 @@
 ## Breaking changes
 
 * Now windowed functions follow the `purrr` style exactly (#35): 
-  + `slide()`, `tile()`, `stretch()` return an output of lists instead of numerics before.
-  + added other type-stable variants like `slide_dbl()`, `slide_int()`, `slide_chr()`, `slide_lgl()`.
+  + `slide()`, `tile()`, `stretch()` only accept atomic vector, and return lists instead of numerics before.
+  + added `slide2()`, `pslide()` to map over multiple inputs simultaneously. 
+  + `lside()`, `ltile()`, `lstretch()` take lists as input.
+  + added other type-stable variants including `slide_dbl()`, `slide_int()`, `slide_chr()`, `slide_lgl()`.
 * Renamed `x` to `...` in windowed functions, `slider()`, `tiler()` and `stretcher()`, to take multiple inputs.
 
 ## New features